added multiply and tried some templates ...?? Is it possible to define templates and styles in one component and use it in the other???

+++ MEF/Component.Toolkit/WorkspaceViewModel.cs
@@ -12,8 +12,8 @@ namespace Component.Toolkit
 
         }
 
-        private int result;
-        public int Result
+        private double result;
+        public double Result
         {
             get { return result; }
             set
